requirements.txt

fastapi
pandas
uvicorn
scikit-learn
shap
pyyaml
numpy
sqlalchemy
psycopg2-binary